requirements.txt
kafka-python==2.0.1
emoji
tweepy==3.10.0